requirements.txt
aiogram==3.22.0
sqlalchemy==2.0.44
asyncpg==0.31.0
pydantic==2.11.10
pydantic-settings==2.12.0
fastapi==0.122.0
uvicorn==0.38.0
python-dotenv==1.2.1